Simplify further by moving epsilon estimation to common funciton
